+++ discord-ec2-manager/create/go.mod
@@ -1,0 +1,16 @@
+module discord-ec2-manager/discord-ec2-manager/create
+
+go 1.17
+
+require (
+	github.com/aws/aws-sdk-go-v2 v1.11.0
+	github.com/aws/aws-sdk-go-v2/service/ec2 v1.21.0
+)
+
+require (
+	github.com/aws/aws-sdk-go-v2/internal/configsources v1.1.0 // indirect
+	github.com/aws/aws-sdk-go-v2/internal/endpoints/v2 v2.0.0 // indirect
+	github.com/aws/aws-sdk-go-v2/service/internal/presigned-url v1.5.0 // indirect
+	github.com/aws/smithy-go v1.9.0 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+)
