+++ discord-ec2-manager/go.mod
@@ -3,8 +3,7 @@ module discord-ec2-manager/discord-ec2-manager
 go 1.16
 
 require (
-	github.com/aws/aws-sdk-go-v2 v1.4.0
-	github.com/aws/aws-sdk-go-v2/config v1.1.7
-	github.com/aws/aws-sdk-go-v2/service/ec2 v1.6.0
-	github.com/bwmarrin/discordgo v0.23.2
+	github.com/aws/aws-sdk-go-v2/config v1.1.7 // indirect
+	github.com/aws/aws-sdk-go-v2/service/ec2 v1.6.0 // indirect
+	github.com/bwmarrin/discordgo v0.23.2 // indirect
 )
